requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
brutils==2.2.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
inflect==7.5.0
Mako==1.3.10
MarkupSafe==3.0.2
more-itertools==10.7.0
psycopg==3.2.9
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
sqlacodegen==3.1.0
SQLAlchemy==2.0.41
starlette==0.47.2
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
